remove_network_ovirt: correct API call, delete the expected network
The original code produced the following error when executed:

    Traceback (most recent call last):
      File "/usr/lib/python2.7/site-packages/celery/app/trace.py", line 375, in trace_task
        R = retval = fun(*args, **kwargs)
      File "/usr/lib/python2.7/site-packages/celery/app/trace.py", line 632, in __protected_call__
        return self.run(*args, **kwargs)
      File "/usr/lib/python2.7/site-packages/socrates_api/tasks.py", line 2744, in remove_network
        return remove_network_ovirt(asset, network)
      File "/usr/lib/python2.7/site-packages/socrates_api/tasks.py", line 1870, in remove_network_ovirt
        api.system_service().vnic_profiles_service().vnic_profile_service(vnic_profile[0].id).remove()
    AttributeError: 'VnicProfilesService' object has no attribute 'vnic_profile_service'

After renaming the `vnic_profile_service` to `profile_service` it was discovered that the `.list(name=name)` filtering didn't work as intended.  ovirt-engine-sdk-python 4.2.9 appears to warn about an unsupported `name` parameter whereas all the 4.3 releases accept it and return full, non-filtered lists.  That usually caused the wrong vNIC profile and network to be deleted.

The `break` statements are purely a protective measure.